**Without using eval():
Fixed the "sanitizing" process for html pages when using Python 3.0 so that the css info is not removed.  We can now browse the NEW official Python 3.0 tutorial properly :-) (only using 3.0a1 - not 3.0a2)

This fixes issue 78.

Note that the OLD tutorial raises a different error as ElementTree trees are using dicts (under Python 3.0) instead of list and this results in an exception when trying to remove unwanted elements/attributes.