Single commit following a long programming session with no Internet access.
Significant improvement over the old "no markup" option (including some unit tests for it,
and a html page explaining the changes).

Changed ":" to "_" for uids like Johannes had done in the better_javascript branch,
in preparation for using jquery.

Fixed a problem with safari when we loaded a default page (index.htm[l]) from a directory
without specifying it fully; safari used to display the html content as a text file
with no processing of the markup.

Fixed a minor bug with doc_help tooltip.

Renamed a few files to better reflect their changed focus.