Started work to define a Crunchy style for Pygments.
Included new way to add new user preferences (as properties): these may be added by a plugin, so that it is not required to edit configuration.py.  We should, whenever possible, move the preferences to the plugin where they belong so as to make them more self-contained *and* de-clutter configuration.py.